Change all float to double.

+++ src/knobab/KnowledgeBase.cpp
@@ -373,12 +373,12 @@ std::pair<const uint32_t, const uint32_t> KnowledgeBase::resolveCountingData(con
     return count_table.resolve_primary_index(mappedVal);
 }
 
-std::unordered_map<uint32_t, float> KnowledgeBase::exists(const std::pair<const uint32_t, const uint32_t>& indexes, const uint16_t& amount) const {
-    std::unordered_map<uint32_t, float> foundElems = std::unordered_map<uint32_t, float>();
+std::unordered_map<uint32_t, double> KnowledgeBase::exists(const std::pair<const uint32_t, const uint32_t>& indexes, const uint16_t& amount) const {
+    std::unordered_map<uint32_t, double> foundElems = std::unordered_map<uint32_t, double>();
 
     for (auto it = count_table.table.begin() + indexes.first; it != count_table.table.begin() + indexes.second + 1; ++it) {
         uint16_t approxConstant = act_table_by_act_id.getTraceLength(it->id.parts.trace_id) / 2;
-        float satisfiability = getSatisifiabilityBetweenValues(amount, it->id.parts.event_id, approxConstant);
+        double satisfiability = getSatisifiabilityBetweenValues(amount, it->id.parts.event_id, approxConstant);
         foundElems.emplace(it->id.parts.trace_id, satisfiability);
     }
